pyproject.toml
[project]
name = "xai-hackathon-epf-methodology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=2.19.1",
    "numpy==1.24.3",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
]
